\name{hayai-package}
\alias{hayai-package}
\alias{hayai}
\docType{package}
\title{
\packageTitle{hayai}
}
\description{
\packageDescription{hayai}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hayai}
\packageIndices{hayai}
### To install database (use this function only first time you use Hayai-annotation Plants)
# db()
### After install database, call Hayai-Annotation Plants
# hayai()
}
\author{
\packageAuthor{hayai}

Maintainer: \packageMaintainer{hayai}
}
\references{
Hayai-Annotation Plants: an ultra-fast and comprehensive functional gene annotation system in plants, Andrea Ghelfi, Kenta Shirasawa, Hideki Hirakawa, Sachiko Isobe, Bioinformatics, btz380, https://doi.org/10.1093/bioinformatics/btz380
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
### To install database (use it only first time you use Hayai-annotation Plants)
# db()
### After install database, call Hayai-Annotation Plants
hayai()
}
